fix(flight-control) remove bugs
