+++ platformio.ini
@@ -15,8 +15,6 @@ framework = arduino
 monitor_speed = 115200
 board_build.mcu = esp32
 board_build.f_cpu = 240000000L
-lib_ldf_mode = chain+
-build_flags = -I include
 lib_deps = 
 	adafruit/Adafruit BMP085 Library@^1.2.1
 	adafruit/Adafruit MPU6050@^2.0.6
@@ -35,8 +33,6 @@ board_build.mcu = esp32
 board_build.f_cpu = 240000000L
 framework = arduino
 monitor_speed = 115200
-lib_ldf_mode = chain+
-build_flags = -I include
 lib_deps = 
 	adafruit/Adafruit BMP085 Library@^1.2.1
 	adafruit/Adafruit MPU6050@^2.0.6
